Revert "Update(Java): 10 add the validations to the CalculadoraUTN class"
This reverts commit bf9f3e6b446e93d4d6907361e48c5f6567713b30.

+++ Java/Tercer Semestre/Leccion09/CalculadoraUTN/src/CalculadoraUTN.java
@@ -41,11 +41,11 @@ public class CalculadoraUTN {
 
     private static void ejecutarOperacion(int operacion, Scanner entrada){
         System.out.print("Digite el Valor para el operando1: ");
-        var operando1 = Double.parseDouble(entrada.nextLine());
+        var operando1 = Integer.parseInt(entrada.nextLine());
         System.out.print("Digite el Valor para el operando2: ");
-        var operando2 = Double.parseDouble(entrada.nextLine());
+        var operando2 = Integer.parseInt(entrada.nextLine());
 
-        Double resultado;
+        int resultado;
         switch (operacion) {
             case 1 -> {//Suma
                 resultado = operando1 + operando2;
